requirements.txt
absl-py==0.11.0
aioredis==1.3.1
alembic==1.5.4
amqp==5.0.9
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
audioread==2.1.9
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.4
av==8.1.0
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bleach==4.1.0
cachetools==4.2.1
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.2.3
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.3.0
chardet==4.0.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cliff==3.6.0
cmaes==0.8.0
cmd2==1.5.0
colorama==0.4.4
colorlog==4.7.2
configparser==5.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect2==0.4.1
cycler==0.10.0
daphne==3.0.2
dataclasses==0.6
decorator==4.4.2
defusedxml==0.7.1
Django==3.1.5
django-celery-results==2.2.0
django-cors-headers==3.6.0
django-environ==0.4.5
django-extensions==3.1.3
django-hosts==4.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
dlib==19.21.1
docopt==0.6.2
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gizeh==0.1.11
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.7.2
google-cloud-texttospeech==2.2.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.0.4
gyp==0.1
hiredis==2.0.0
httpagentparser==1.9.1
httplib2==0.18.1
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
incremental==21.3.0
inflection==0.5.1
install==1.3.4
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.4.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
kombu==5.2.3
libcst==0.3.16
librosa==0.8.0
llvmlite==0.35.0
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
moviepy==1.0.3
mpmath==1.1.0
msgpack==1.0.2
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
nltk==3.6.2
notebook==6.4.4
numba==0.52.0
numpy==1.19.3
oauth2client==4.1.3
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-contrib-python==4.5.1.48
opencv-python==4.4.0.46
optuna==2.5.0
packaging==20.8
pandas==1.2.1
pandocfilters==1.5.0
parso==0.7.1
pbr==5.5.1
pdf2image==1.14.0
pescador==2.1.0
pexpect==4.8.0
phonenumbers==8.12.15
pickleshare==0.7.5
pika==1.2.0
Pillow==8.1.0
pkgconfig==1.5.5
pooch==1.3.0
prettytable==0.7.2
proglog==0.1.9
prometheus-client==0.11.0
prompt-toolkit==3.0.8
proto-plus==1.13.0
protobuf==3.14.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyannote.core==4.1
pyannote.database==4.0.4
pyannote.metrics==3.0.1
pyannote.pipeline==1.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDrive==1.3.1
pydub==0.24.1
Pygments==2.7.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.1
pypexels==1.0.0rc1
pyrsistent==0.18.0
python-dateutil==2.8.1
python-editor==1.0.4
python-magic==0.4.18
python-pixabay==4.2
pytz==2021.3
pyunsplash==1.0.0rc2
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==5.1.1
QtPy==1.11.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.5.4
selenium==3.141.0
Send2Trash==1.8.0
service-identity==21.1.0
shellingham==1.4.0
simplejson==3.17.2
six==1.15.0
sortedcollections==2.1.0
sortedcontainers==2.3.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
stevedore==3.3.0
stripe==2.67.0
sympy==1.7.1
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2021.2.1
tinycss2==1.1.0
toml==0.10.2
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
typer==0.3.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
Wand==0.6.5
wcwidth==0.2.5
webencodings==0.5.1
webp==0.1.3
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zope.interface==5.4.0
